Resolve canonical WebFinger subjects of remote actors (#349)
* webfinger: Add ability to resolve canonical `subject`s

* Make confirmation query even in case of same-origin redirects

* type: Make `aliases` and `links` properties of JRD resource optional

These properties are OPTIONAL according to RFC 7033.

cf. <https://datatracker.ietf.org/doc/html/rfc7033#section-4.4>

* webfinger: Case-sensitively compare domain of JRD subject

Although RFC 3986 Section 3.2.2. specifies that the host subcomponent of
authority is case-insensitive, an odd server might return different
responses for different casing, so it might be safer to retry the query
with the casing of the  returned JRD.

Also, the implementation is significantly simpler this way.

* webfinger: Construct full URI for query in every iteration

I introduced `uri::Parts` to be stingy with heap allocations that occur
while updating the URI, but the change has increased complexity of the
    code, and in fact it did two allocations per an update in common
    cases so it's not certain that it was even a valid optimisation.

* Resolve WebFinger resource after fetching ActivityPub actor

* Fix invalid `preferredUsername` in test

* Use mocked HTTP client for WebFinger in test for `resolve::post`